Feature/my trip short

+++ src/main/java/com/example/bami/short_travel/dto/SaveShortTravelDTO.java
@@ -14,6 +14,6 @@ public class SaveShortTravelDTO {
     private List<RecommendationDTO> recommendations;
     private String startDate;
     private String endDate;
-    private Double latitude;
-    private Double longitude;
+    private float latitude;
+    private float longitude;
 }
